// chpqa-app/src/app/api-services/chpqa-api/generated/model/requestScheme.ts
/**
 * WebApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RequestEconomicSubSector } from './requestEconomicSubSector';
import { RequestSchemeLocationDetails } from './requestSchemeLocationDetails';
import { Organisation } from './organisation';
import { RequestEconomicSector } from './requestEconomicSector';
import { Person } from './person';


export interface RequestScheme { 
    name: string;
    reference?: string | null;
    createdon?: string | null;
    status: number;
    statusText?: string | null;
    company: Organisation;
    responsiblePerson: Person;
    site: RequestSchemeLocationDetails;
    econSector?: RequestEconomicSector;
    econSubSector?: RequestEconomicSubSector;
    sicCodeId?: string | null;
}