// chpqa-app/src/app/api-services/chpqa-api/generated/model/requestSchemeLocationDetails.ts
/**
 * WebApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RequestSiteContact } from './requestSiteContact';


export interface RequestSchemeLocationDetails { 
    id?: string | null;
    name?: string | null;
    address1: string;
    address2?: string | null;
    postcode: string;
    town: string;
    county?: string | null;
    instructions?: string | null;
    areYouTheSiteContactPerson?: boolean | null;
    contactPerson?: RequestSiteContact;
}